refactor: move localstorageMock
