+++ mocks/localStorageMock.js
